web_interface/requirements.txt
Django>=4.2.0
channels>=4.0.0
redis>=5.0.0
daphne>=4.0.0
django-redis>=5.3.0
aioredis>=2.0.1
channels-redis>=4.1.0
pyserial>=3.5
python-dotenv>=1.0.0 